+++ src/main/java/ispp/project/dondesiempre/repositories/stores/StoreRepository.java
@@ -2,12 +2,15 @@ package ispp.project.dondesiempre.repositories.stores;
 
 import ispp.project.dondesiempre.models.stores.Store;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 public interface StoreRepository extends JpaRepository<Store, UUID> {
+
+  Optional<Store> findByUserId(UUID userId);
 
   @Query(
       value =
