+++ src/main/java/ispp/project/dondesiempre/services/UserService.java
@@ -4,8 +4,9 @@ import ispp.project.dondesiempre.exceptions.ResourceNotFoundException;
 import ispp.project.dondesiempre.exceptions.UnauthorizedException;
 import ispp.project.dondesiempre.models.Client;
 import ispp.project.dondesiempre.models.User;
+import ispp.project.dondesiempre.models.stores.Store;
 import ispp.project.dondesiempre.repositories.ClientRepository;
-import ispp.project.dondesiempre.repositories.UserRepository;
+import ispp.project.dondesiempre.repositories.stores.StoreRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationContext;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -19,18 +20,10 @@ public class UserService {
   /** Email used for the seed store owner. Referenced by DataSeeder. */
   public static final String SEED_USER_EMAIL = "owner@laboutique.es";
 
-  private final UserRepository userRepository;
   private final ClientRepository clientRepository;
+  private final StoreRepository storeRepository;
   private final PasswordEncoder passwordEncoder;
   private final ApplicationContext applicationContext;
-
-  @Transactional(rollbackFor = Exception.class)
-  public User register(String email, String password) {
-    User user = new User();
-    user.setEmail(email);
-    user.setPassword(passwordEncoder.encode(password));
-    return userRepository.save(user);
-  }
 
   public boolean checkPassword(User user, String rawPassword) {
     return passwordEncoder.matches(rawPassword, user.getPassword());
@@ -44,5 +37,15 @@ public class UserService {
     return clientRepository
         .findByUserId(currentUser.getId())
         .orElseThrow(() -> new ResourceNotFoundException("Current client not found."));
+  }
+
+  @Transactional(
+      readOnly = true,
+      rollbackFor = {ResourceNotFoundException.class, UnauthorizedException.class})
+  public Store getCurrentStore() throws ResourceNotFoundException, UnauthorizedException {
+    User currentUser = applicationContext.getBean(AuthService.class).getCurrentUser();
+    return storeRepository
+        .findByUserId(currentUser.getId())
+        .orElseThrow(() -> new ResourceNotFoundException("Current store not found."));
   }
 }
